requirements.txt
aiogram>=3.7.0
python-dotenv>=1.0.0

fastapi>=0.111.0
uvicorn[standard]>=0.30.0

SQLAlchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.20.0
greenlet>=3.0.0

pydantic>=2.0.0
httpx>=0.27.0

Jinja2>=3.1.0